+++ cores/arduino/HardwareSerial.h
@@ -42,7 +42,7 @@ class HardwareSerial : public Stream
 {
   public:
     virtual void begin(unsigned long);
-    virtual void begin(unsigned long baudrate, uint16_t config);
+    virtual void begin(unsigned long baudrate, uint32_t config);
     virtual void end();
     virtual int available(void) = 0;
     virtual int peek(void) = 0;
